requirements.txt
async-generator==1.10
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
cryptography==37.0.2
h11==0.13.0
idna==3.3
multivolumefile==0.2.3
numpy==1.22.3
outcome==1.1.0
pandas==1.4.2
patool==1.12
psutil==5.9.0
py7zr==0.18.5
pybcj==0.5.2
pycparser==2.21
pycryptodomex==3.14.1
pyOpenSSL==22.0.0
pyppmd==0.18.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzstd==0.15.2
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
texttable==1.6.4
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
zipfile-deflate64==0.2.0
